Fix URLs for non-default combinations
Products in listings not always represent default combinations. For example default combination is RED, and it costs 10 EUR. If displayed product represents default combination, $product.url is same as $product.canonical_url, it does not have an attribute anchor, so it is OK. 
But if displayed product represents another combination, let's say color GREEN, it has another price, and it should have another link, with attribute anchor. When customer clicks on that product, he should see the GREEN combination when product page is loaded.